third_party/WebKit/LayoutTests/fast/dom/shadow/create-shadow-root-with-parameter-expected.txt

Create open shadow root.
PASS [object ShadowRoot] is non-null.
Create closed shadow root.
PASS host2.createShadowRoot({mode: 'closed'}) threw exception NotSupportedError: Failed to execute 'createShadowRoot' on 'Element': Closed shadow root is not implemented yet..
Create shadow root with empty parameter.
PASS [object ShadowRoot] is non-null.
Create shadow root whose mode is neither open nor closed.
PASS host4.createShadowRoot({mode: 'illegal'}) threw exception TypeError: Failed to execute 'createShadowRoot' on 'Element': The provided value 'illegal' is not a valid enum value of type ShadowRootMode..
Create open shadow root with shadow-dom.js utility.
PASS [object ShadowRoot] is non-null.
PASS successfullyParsed is true

TEST COMPLETE
